Extract grouping code to dedicated module
